Solution to Calculate Average Temperature of a Week

+++ Calculate Average Temperature of a Week@@67877d61dc107486ad5c274d/code.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
 
 struct week{
-    float temp;
+    int temp;
 };
 
 int main() {
     int n=7;
     struct week e[7];
-    float totalTemp=0;
+    int totalTemp=0;
     for(int i=0;i<n;i++){
         scanf("%f",&e[i].temp);
         totalTemp += e[i].temp;
